Make create new text tag button into individual class.
